chore: Moves ElasticPress and WordPress AI compatibility classes do integrations folder.
