+++ src/views/admin/class-tainacan-admin.php
@@ -290,7 +290,7 @@ class Admin extends Pages {
 		 * Must not be called from get_admin_js_localization_params() because that runs on 'init' priority 11,
 		 * before the AI plugin registers abilities on 'init' at 15.
 		 */
-		$plugin_settings['ai_alt_text_generation_available'] = \Tainacan\WordPress_AI::get_instance()->is_alt_text_generation_available();
+		$plugin_settings['ai_alt_text_generation_available'] = \Tainacan\Integrations\WordPress_AI::get_instance()->is_alt_text_generation_available();
 		wp_localize_script( 'tainacan-pages-common-scripts', 'tainacan_plugin', $plugin_settings );
 		
 		wp_enqueue_media();
